+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Enhancements:
 * Add support for Logs visualization, fix https://github.com/Altinity/clickhouse-grafana/issues/331, thanks @Fiery-Fenix and @pixelsquared 
 * Add $conditionalTest to editor auto-complete
+* Add support $__searchFilter to template variable queries, fix https://github.com/Altinity/clickhouse-grafana/issues/354
 
 ## Fixes:
 * allow Nullable types in alert label name in backend part, fix https://github.com/Altinity/clickhouse-grafana/issues/405